chore: new explorer page
