+++ packages/apps/config/gql.ts
@@ -19,39 +19,9 @@ export const STATISTICS_QUERY = `
   }
 `;
 
-export const HISTORY_RECORDS = `
-  query historyRecords($row: Int!, $page: Int!, $sender: String, $recipient: String) {
-    historyRecords(row: $row, page: $page, sender: $sender, recipient: $recipient) {
-      total
-      records {
-        sendAmount
-        recvAmount
-        bridge
-        endTime
-        fee
-        feeToken
-        fromChain
-        id
-        nonce
-        messageNonce
-        recipient
-        requestTxHash
-        responseTxHash
-        reason
-        result
-        sender
-        startTime
-        toChain
-        sendToken
-        recvToken
-      }
-    }
-  }
-`;
-
 export const HISTORY_RECORDS_IN_RESULTS = `
-  query historyRecords($row: Int!, $page: Int!, $sender: String, $results: [Int]) {
-    historyRecords(row: $row, page: $page, sender: $sender, results: $results) {
+  query historyRecords($row: Int!, $page: Int!, $sender: String, $recipient: String, $results: [Int]) {
+    historyRecords(row: $row, page: $page, sender: $sender, recipient: $recipient, results: $results) {
       total
       records {
         sendAmount
