+++ packages/apps/styles/index.scss
@@ -54,21 +54,38 @@ body {
   }
 }
 
+.explorer-tabs > .ant-tabs-nav {
+  margin-bottom: 0 !important;
+}
+
 .explorer-table {
+  $spacing: 10px;
+
+  .ant-table table {
+    border-spacing: 0 $spacing;
+  }
+
   .ant-table.ant-table-small {
-    background: rgba($bg-dark, 0.5);
+    background: transparent;
 
     .ant-table-thead > tr > th {
-      padding-top: 16px;
-      padding-bottom: 16px;
+      padding: 16px 15px;
       background: $bg-dark;
       box-shadow: 0 4px 20px rgba(0, 0, 0, 0.1);
       backdrop-filter: blur(50px);
     }
 
     .ant-table-tbody > tr > td {
-      padding-top: 10px;
-      padding-bottom: 10px;
+      padding: 10px 15px;
+      border-top: 1px solid #303030;
+
+      &:first-child {
+        border-left: 1px solid #303030;
+      }
+
+      &:last-child {
+        border-right: 1px solid #303030;
+      }
     }
   }
 }
